Repair waypoints.py in place to use pfaedle (one pipeline); drop parallel script/CI
Per review: rather than a parallel generator + separate CI that only uploads an
artifact (which consumers wouldn't discover, and creates a two-version ambiguity),
repair the EXISTING pipeline so gh-pages serves the map-matched data directly.

- waypoints.py now map-matches the official GTFS to OSM via pfaedle, covering buses
  AND green minibuses (route_type 3) in one pass; rail/tram/ferry static files unchanged.
- removed waypoints_pfaedle.py and the parallel crawl-pfaedle.yml workflow.
- dropped the geopandas stack from requirements (stdlib + Docker only).
- autopep8-clean so the Format check passes.

Verified end-to-end: 3520 route-directions, pfaedle matched all 3520, exit 0.
