--- .github/workflows/crawl-pfaedle.yml
@@ -1,27 +1,0 @@
-name: Data Fetching (pfaedle / OSM map-match, RFC)
-
-# Runs the pfaedle-based generator and uploads the result as an artifact for inspection.
-# Does NOT deploy to gh-pages, so it never touches live data. Manual trigger + this RFC
-# branch (the branch trigger is here so reviewers can see it go green; drop it before merge).
-on:
-  workflow_dispatch:
-  push:
-    branches: [pfaedle-waypoints]
-
-jobs:
-  Fetch-Route-Data-pfaedle:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check out repository code
-        uses: actions/checkout@v4
-      - name: Setup Python environment
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.12'
-      - name: Generate waypoints via pfaedle (Docker is preinstalled on the runner)
-        run: python waypoints_pfaedle.py
-      - name: Upload waypoints artifact
-        uses: actions/upload-artifact@v4
-        with:
-          name: waypoints-pfaedle
-          path: waypoints
