+++ README.md
@@ -2,7 +2,9 @@
 
 ![Data fetching status](https://github.com/hkbus/route-waypoints/actions/workflows/crawl.yml/badge.svg) 
 
-This project is to fetch the waypoints from [CSDI](https://portal.csdi.gov.hk/geoportal/#metadataInfoPanel). It is daily synced to the sources and launched in gh-pages.
+This project generates route waypoints by **map-matching the official GTFS stop sequence to OpenStreetMap** with [`pfaedle`](https://github.com/ad-freiburg/pfaedle). It is daily synced to the sources and launched in gh-pages.
+
+> **Why not CSDI?** The [CSDI](https://portal.csdi.gov.hk/geoportal/#metadataInfoPanel) line is not GPS-observed — it is itself routed between stops (unrelated routes share kilometres of byte-identical vertices), so it inherits a router's mistakes: spurious multi-km loops at some termini and long detours where a turn was skipped (e.g. route 15 never walks the correct road). Re-deriving each shape from the stops with a real map-matcher fixes every class at once and lands within ~1% of the operators' own published lines. Covers buses **and** green minibuses (GTFS `route_type` 3) in one pass; rail/tram/ferry keep their curated static files.
 
 During the crawling, it will minified the result by truncating to 5 decimal places, (i.e., ±1m), and minified the json by cleaning useless space characters. Also, as the data is provided statically by `gh-pages`, the data transfer supports `Content-Encoding: gzip` for largely preserving your bandwidth.
 
@@ -19,36 +21,14 @@ Example link: (https://hkbus.github.io/route-waypoints/1001-O.json)
 ### Usage
 Daily fetched GeoJSONs are in [gh-pages](https://github.com/hkbus/route-waypoints/tree/gh-pages).
 
-### Installation
-
-To install the dependencies,
-```
-pip install -r ./crawling/requirements.txt
-```
-
 ### Data Fetching
 
-To fetch data, run the followings,
+`waypoints.py` needs only the standard library plus **Docker** (pfaedle runs from its official
+image; Docker is preinstalled on GitHub's ubuntu runners). To fetch data:
 ```
 python ./waypoints.py
 ```
-
-## Experimental: waypoints from OpenStreetMap map-matching (`waypoints_pfaedle.py`)
-
-The CSDI line is **not GPS-observed** — it is itself routed between stops, so it inherits a
-router's mistakes: spurious multi-km loops at some termini and long detours where a turn was
-skipped. `waypoints_pfaedle.py` instead **re-derives** each shape from the official GTFS stop
-sequence by map-matching to OpenStreetMap with [`pfaedle`](https://github.com/ad-freiburg/pfaedle),
-and writes the **same** `{ROUTE_ID}-{O|I}.json` format. It fixes every defect class at once and,
-on a spot-check, lands within ~1% of the operators' own published route lines.
-
-```
-python waypoints_pfaedle.py          # needs Docker (pfaedle runs from the official image)
-```
-
-Runs the whole franchised-bus network in about a minute (OSM graph load + map-match). This is an
-RFC — the manual `Data Fetching (pfaedle …)` workflow uploads the output as an artifact and does
-**not** deploy to `gh-pages`. See the pull request for the evidence.
+It runs the whole road-transport network in about a minute (OSM graph load + map-match).
 
 ## Citing
